fix: conversion warnings

+++ src/MenuButton.cpp
@@ -13,8 +13,8 @@ void MenuButton::updateSurface() {
 	if (m_surface) SDL_DestroySurface(m_surface);
 
 	m_surface = TTF_RenderText_Solid(m_font, m_text.c_str(), m_text.length(), m_isSelected ? m_selectedColor : m_color);
-	m_width = m_surface->w;
-	m_height = m_surface->h;
+	m_width = static_cast<float>(m_surface->w);
+	m_height = static_cast<float>(m_surface->h);
 }
 
 void MenuButton::render(SDL_Renderer* renderer, const SDL_FRect* dest) {
